# Another performance improvement on the Formatter by using Kernel#sprintf over String#%